# api/openapi/openapi-3.0.yaml
openapi: 3.0.3
info:
  title: LinguaFlow Web Service API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /mode:
    get:
      operationId: GetMode
      summary: 获取运行模式
      security: []
      description: 返回当前服务的运行模式（local 单用户本地模式 / server 多租户服务端模式），供前端决定是否跳过登录页。无需鉴权。
      responses:
        '200':
          description: 当前模式
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeResponse'
        default:
          $ref: '#/components/responses/Problem'
  /auth/register:
    post:
      operationId: RegisterAuth
      summary: 注册用户并签发会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        default:
          $ref: '#/components/responses/Problem'
  /auth/login:
    post:
      operationId: LoginAuth
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        default:
          $ref: '#/components/responses/Problem'
  /auth/refresh:
    post:
      operationId: RefreshAuth
      summary: 刷新访问令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        default:
          $ref: '#/components/responses/Problem'
  /auth/logout:
    post:
      operationId: LogoutAuth
      summary: 登出并撤销刷新令牌
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: 已登出
        default:
          $ref: '#/components/responses/Problem'
  /users/me:
    get:
      operationId: GetCurrentUser
      summary: 获取当前用户
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: UpdateCurrentUser
      summary: 更新当前用户资料
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentUserRequest'
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/Problem'
  /users/me/password:
    put:
      operationId: ChangeCurrentUserPassword
      summary: 修改当前用户密码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: 密码已更新
        default:
          $ref: '#/components/responses/Problem'
  /orgs:
    get:
      operationId: ListOrganizations
      summary: 列出当前用户所属组织
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 组织列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: CreateOrganization
      summary: 创建组织
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        '201':
          description: 已创建组织
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    get:
      operationId: GetOrganization
      summary: 获取组织详情
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 组织详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: UpdateOrganization
      summary: 更新组织信息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRequest'
      responses:
        '200':
          description: 更新后的组织
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}/members:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    get:
      operationId: ListOrganizationMembers
      summary: 获取组织成员列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成员列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMemberListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: AddOrganizationMember
      summary: 添加组织成员
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOrganizationMemberRequest'
      responses:
        '201':
          description: 已添加成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMember'
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/UserId'
    put:
      operationId: UpdateOrganizationMember
      summary: 修改组织成员角色
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationMemberRequest'
      responses:
        '200':
          description: 更新后的成员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMember'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: DeleteOrganizationMember
      summary: 移除组织成员
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 已移除成员
        default:
          $ref: '#/components/responses/Problem'
  /backends:
    post:
      operationId: CreateUserBackend
      summary: 创建个人后端
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBackendRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Backend'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: ListUserBackends
      summary: 列出个人后端
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 后端列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /backends/{backendId}:
    parameters:
      - $ref: '#/components/parameters/BackendId'
    put:
      operationId: UpdateUserBackend
      summary: 更新个人后端
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBackendRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Backend'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: DeleteUserBackend
      summary: 删除个人后端
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}/backends:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    post:
      operationId: CreateOrgBackend
      summary: 创建组织后端
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBackendRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Backend'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: ListOrgBackends
      summary: 列出组织后端
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 后端列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackendListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}/backends/{backendId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/BackendId'
    put:
      operationId: UpdateOrgBackend
      summary: 更新组织后端
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBackendRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Backend'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: DeleteOrgBackend
      summary: 删除组织后端
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        default:
          $ref: '#/components/responses/Problem'
  /projects:
    post:
      operationId: CreateProject
      summary: 创建项目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: ListProjects
      summary: 列出项目
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: GetProject
      summary: 获取项目详情
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    put:
      operationId: UpdateProject
      summary: 更新项目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: DeleteProject
      summary: 删除项目
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        default:
          $ref: '#/components/responses/Problem'
  /orgs/{orgId}/projects:
    parameters:
      - $ref: '#/components/parameters/OrgId'
    post:
      operationId: CreateOrgProject
      summary: 创建组织项目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        default:
          $ref: '#/components/responses/Problem'
    get:
      operationId: ListOrgProjects
      summary: 列出组织项目
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /admin/stats:
    get:
      operationId: AdminGetStats
      summary: 获取全局统计
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 全局统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStats'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Problem:
      description: 错误响应
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    ModeResponse:
      type: object
      required:
        - mode
      properties:
        mode:
          type: string
          enum:
            - local
            - server
          x-enum-varnames:
            - ModeLocal
            - ModeServer
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
...
    AuthSession:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_at
        - refresh_expires_at
        - user
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_at:
          type: string
          format: date-time
        refresh_expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...